Finished Assignment1.java
My program prints my first and last initials (AC) in large block letters just as instructed in the markdown file. Then, my code asks the user to input a number in Fahrenheit. My program converts the value to Celsius and saves it.  Then, my program takes input from the user in the form of a 5-character string. It saves the string in reverse without the first and last characters. Them, my program generates and saves a random number between 32 and 16384 (inclusive). Finally, my program combines the results of the 2nd, 3rd, and 4th steps as per the markdown file and prints it to the screen. Then my program closes the scanner and my program is finished.